requirements.txt
git+https://bitbucket.org/mapdes/fiat.git#egg=fiat
hg+https://bitbucket.org/khinsen/scientificpython
pytest-pep8
pymbolic
numpy>=1.6.1
sympy
